lambda/requirements.txt
awslabs-mcp-lambda-handler==0.1.8
boto3==1.40.10
botocore==1.40.10
jmespath==1.0.1
python-dateutil==2.9.0.post0
s3transfer==0.13.1
six==1.17.0
urllib3==2.5.0
